dev(kernel) : refactor kernel
