+++ src/drivers/tty/tty.c
@@ -1,8 +1,8 @@
 #include <libft.h>
 
-#include "acpi.h"
 #include "internal/tty.h"
 #include "keyboard.h"
+#include <acpi.h>
 #include <vga.h>
 
 static bool ft_strequ(const char *s1, const char *s2)
