+++ src/idt.c
@@ -1,11 +1,11 @@
 #include <libft.h>
 
-#include "acpi.h"
 #include "idt.h"
 #include "io.h"
 #include "keyboard.h"
 #include "panic.h"
 #include "vga.h"
+#include <acpi.h>
 
 #define PIC1         0x20 // IO base address for master PIC
 #define PIC2         0xA0 // IO base address for slave PIC
