+++ src/kernel/internal/panic.h
@@ -1,10 +1,12 @@
 #pragma once
 
 // ============================================================================
-// INCLUDES
+// IMCLUDES
 // ============================================================================
 
-#include <kernel/mb2_info.h>
+#include "panic.h"
+#include <acpi.h>
+#include <drivers/vga.h>
 #include <types.h>
 
 // ============================================================================
@@ -12,8 +14,6 @@
 // ============================================================================
 
 // Defines
-
-enum allocator_state { ACTIVE = 0, FROZEN };
 
 // Macros
 
@@ -23,34 +23,17 @@ enum allocator_state { ACTIVE = 0, FROZEN };
 
 // Enums
 
-enum mem_type { FREE_MEMORY = 0, RESERVED_MEMORY, HOLES_MEMORY };
-
 // Structures
 
-struct region_s {
-	uintptr_t start;
-	uintptr_t end;
-};
-
-struct boot_allocator;
-
 // Typedefs
-
-typedef struct region_s       region_t;
-typedef struct boot_allocator boot_allocator_t;
 
 // ============================================================================
 // VARIABLES GLOBALES
 // ============================================================================
 
 // ============================================================================
-// EXTERNAL APIs
+// INTERNAL APIs
 // ============================================================================
 
-bool      boot_allocator_range_overlaps(uintptr_t start, uintptr_t end, enum mem_type type);
-void     *boot_alloc(uint32_t size);
-void      boot_allocator_freeze(void);
-void      boot_allocator_printer(void);
-void      boot_allocator_init(multiboot_tag_mmap_t *mmap, uint8_t *mmap_end);
-uint32_t  boot_allocator_get_region_count(enum mem_type type);
-region_t *boot_allocator_get_region(enum mem_type type);
+void memory_dump(uint32_t addr_start, uint32_t addr_end);
+void save_stack(void);
