+++ src/kernel/main.c
