+++ src/memory/internal/buddy.h
@@ -4,6 +4,7 @@
 // IMCLUDES
 // ============================================================================
 
+#include <acpi.h>
 #include <memory/buddy.h>
 
 // ============================================================================
